requirements.txt
asgiref==3.4.1
attrs==21.2.0
awsebcli==3.20.0
bcrypt==3.2.0
blessed==1.18.1
botocore==1.20.112
cached-property==1.5.2
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.3
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.6
django-allauth==0.45.0
djangorestframework==3.12.4
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.1
jmespath==0.10.0
jsonschema==3.2.0
ndg-httpsclient==0.5.1
oauthlib==3.1.1
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.3.1
psycopg2-binary==2.9.1
pusher==3.0.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.4
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==5.3.0
zipp==3.5.0
APScheduler==3.7.0
asgiref==3.4.1
boto3==1.18.23
botocore==1.21.23
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.45.0
django-apscheduler==0.6.0
django-storages==1.11.1
django-webpush==0.3.3
djangorestframework==3.12.4
django-imagekit==4.0.2
gunicorn==20.1.0
http-ece==1.1.0
idna==3.2
jmespath==0.10.0
ndg-httpsclient==0.5.1
oauthlib==3.1.1
Pillow==8.3.1
psycopg2-binary==2.9.1
pusher==3.0.0
py-vapid==1.8.2
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
pywebpush==1.9.4
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.6
whitenoise==5.3.0
APScheduler==3.7.0
asgiref==3.4.1
boto3==1.18.25
botocore==1.21.25
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.45.0
django-apscheduler==0.6.0
django-storages==1.11.1
django-webpush==0.3.3
djangorestframework==3.12.4
gunicorn==20.1.0
http-ece==1.1.0
idna==3.2
jmespath==0.10.0
mysqlclient==2.0.3
ndg-httpsclient==0.5.1
oauthlib==3.1.1
Pillow==8.3.1
psycopg2-binary==2.9.1
pusher==3.0.0
py-vapid==1.8.2
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
pywebpush==1.9.4
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
whitenoise==5.3.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
docopt==0.6.2
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-speech==2.6.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
idna==3.2
Js2Py==0.71
libcst==0.3.19
mypy-extensions==0.4.3
packaging==21.0
pipwin==0.5.1
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
tzlocal==2.1
urllib3==1.26.6
APScheduler==3.7.0
asgiref==3.4.1
boto3==1.18.37
botocore==1.21.37
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.45.0
django-appconf==1.0.4
django-apscheduler==0.6.0
django-imagekit==4.0.2
django-pwa==1.0.10
django-storages==1.11.1
django-webpush==0.3.3
djangorestframework==3.12.4
gunicorn==20.1.0
http-ece==1.1.0
idna==3.2
jmespath==0.10.0
mysqlclient==2.0.3
ndg-httpsclient==0.5.1
oauthlib==3.1.1
pilkit==2.0
Pillow==8.3.2
psycopg2-binary==2.9.1
pusher==3.0.0
py-vapid==1.8.2
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
pywebpush==1.9.4
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.6
whitenoise==5.3.0
